+++ Assets/Scripts/State/Core/ShipState.cs
@@ -1,6 +1,6 @@
 ﻿namespace STP.State.Core {
     public class ShipState {
-        public int Hp;
+        public float Hp;
 
         public ShipState(int curHp) {
             Hp = curHp;
